requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django==1.10.6
numpy==1.12.0
pydicom==0.9.9
pytest==3.0.6
pillow==3.0.0